backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
starlette==0.36.3

# Async HTTP client (replaces requests for async)
httpx==0.27.0

# Video processing
av==11.0.0
opencv-python==4.9.0.80
ffmpeg-python==0.2.0
numpy==1.26.4

# Cloud storage
boto3==1.34.67

# Streaming
m3u8==4.0.0

# Time handling
pytz==2024.1

# API validation
pydantic==2.6.4
pydantic_core==2.16.3

# Core dependencies
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
anyio==4.3.0
sniffio==1.3.1
typing_extensions==4.10.0

# REMOVED (no longer needed):
# celery - Replaced with asyncio
# redis - Replaced with in-memory deque
# flower - No Celery = no monitoring
# flask - Not used
# tornado - Celery dependency
# amqp, billiard, kombu, vine - Celery dependencies
